services/edge-node/requirements.txt
# Edge Node Service Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
aiohttp==3.9.1
pillow==10.2.0
numpy==1.26.3

# TODO: Add ML framework for actual model inference
# torch==2.1.2
# onnxruntime==1.16.3
